refactor: crée un fichier TrackingChaptersContext
